refactor(Commons): Refactor the commons package and move kratos related files to specific package
