+++ commons/src/main/java/com/heimdallauth/server/commons/models/kratos/GroupModel.java
@@ -1,4 +1,4 @@
-package com.heimdallauth.server.commons.models;
+package com.heimdallauth.server.commons.models.kratos;
 
 import lombok.*;
