+++ kratos/src/main/java/com/heimdallauth/server/datamanagers/RoleDataManager.java
@@ -1,6 +1,6 @@
 package com.heimdallauth.server.datamanagers;
 
-import com.heimdallauth.server.commons.models.RoleModel;
+import com.heimdallauth.server.commons.models.kratos.RoleModel;
 
 import java.util.List;
 import java.util.Optional;
